feat: a Pickers classification — the one place a component may fetch
The charter said "a widget never fetches," and the one widget that did —
the faction load popover — looked like a violation. It isn't: it is a
control that fetches its OWN options, lazily, so a page never has to
query every loadable faction up front on the off chance the reader opens
the picker. That is a different thing from a widget, and it deserves its
own classification rather than a blanket ban or a blessed exception.

A Picker is a domain control whose job is to let the user choose from a
list its options are its own to load. It fetches only what presenting
those options needs, read-only, torn down when it leaves the screen; it
never mutates, never reads page data; the chosen value leaves through a
callback and the route decides what happens next. The laziness IS the
justification (DD-013 subscription discipline) — a Picker that fetched
eagerly or mutated would just be a widget breaking the rule.

`src/app/pickers/` is a new top-level role (guarded by check:app-layout).
The connected control moves there as `FactionPicker`; the popover shell
stays in the widget as the container that mounts it only while open, so
the picker subscribes on mount because the container already gated it and
opening the popover is the intent signal. The inline shape (defer the
subscription to the control's own open) is documented in the charter but
not built — it lands with its first consumer, not as an unused path.

With the fetch gone, the widget is types-only, so `src/app/widgets/**`
now carries the same fetch-ban as the kit (value @db and convex banned,
type imports allowed, stories/tests excluded). Proven to bite on a value
import and spare a type import.

Behaviour unchanged: the load flow is exercised by faction-lifecycle
e2e (Load existing faction -> Load faction), green. Full suite green:
typecheck, lint, format, knip, check:app-layout, check:css-orphans,
447 app + 279 story tests, 6/6 e2e.
